+++ .gitignore
@@ -1,0 +1,27 @@
+venv/
+
+*.pyc
+__pycache__/
+
+instance/
+
+.pytest_cache/
+.coverage
+htmlcov/
+
+dist/
+build/
+_build/
+html/
+*.egg-info/
+
+.idea/*
+
+concept
+*data.json
+
+emhub/static/images/user-profile*
+*.csv
+
+NOTES.txt
+
