--- cluster/__init__.py
@@ -1,1 +1,0 @@
-from cluster.__version__ import __version__, __title__
