--- cluster/__version__.py
@@ -1,3 +1,0 @@
-__version__ = '0.1'
-
-__title__ = "the folder contains data structure for the cluster jobs/node information"
