+++ emgoat/__init__.py
@@ -1,0 +1,4 @@
+
+__version__ = "0.0.1"
+
+from .config import config
