+++ emgoat/cluster/__init__.py
@@ -1,0 +1,3 @@
+
+from .base import Cluster
+from .lsf import LSFCluster
