+++ emgoat/cluster/lsf/__init__.py
@@ -1,0 +1,2 @@
+
+from .lsf import LSFCluster
