+++ emgoat/cluster/lsf/functions.py
@@ -2,64 +2,21 @@
 """
 This file stores the utility functions 
 """
-import subprocess
-from pwd import getpwnam
 from datetime import datetime
-from util.macros import VERY_BIG_NUMBER, GPU_TYPE
-import csv
 
-def run_command(arglist, user_name=None, timeout=60):
-    """
-    Run a specified command with arguments and return
-    standard output decoded to UTF-8.
+from emgoat.util import (VERY_BIG_NUMBER, JOB_STATUS_PD, JOB_STATUS_RUN, JOB_STATUS_DONE,
+                         convert_str_to_integer, convert_percentage_to_decimal)
 
-    The subprocess can also be launched with another user and group. 
-    In default if the user and group are None, then the command
-    will run in normal way, otherwise it will run under another user's 
-    name. Make sure you have permission to run the commands under
-    the input user!!!
 
-    :param list(str) arglist: List of command and arguments
-    :param user_name(str): run command under another user name
-    :param int timeout: set the timeout to this many seconds(default 120)
-    :returns: Output of the command decoded to utf-8
-    """
-
-    # checking whether the user exists? also get the corresponding uid
-    uid = None
-    if user_name is not None:
-        try:
-            uid = getpwnam(user_name).pw_uid
-        except:
-            info = 'The input user name can not be validated on the OS: {}'.format(user_name)
-            raise KeyError(info)
-
-    # all of output and error directed to the pipe
-    # no standard input needed
-    if user_name is not None:
-        proc = subprocess.Popen(
-            arglist,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            stdin=subprocess.DEVNULL,
-            user=uid
-        )
+def get_job_general_status(status):
+    # get the general status
+    if whether_job_is_pending(status) or whether_job_is_suspending(status):
+        return JOB_STATUS_PD
+    elif whether_job_is_running(status):
+        return JOB_STATUS_RUN
     else:
-        proc = subprocess.Popen(
-            arglist,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            stdin=subprocess.DEVNULL
-        )
+        return JOB_STATUS_DONE
 
-    # run the command, and return the output
-    out, err = proc.communicate(timeout=timeout)
-    if proc.returncode != 0:
-        info = 'Error running command {0}: {1}'.format(arglist, err)
-        raise IOError(info)
-
-    # return
-    return out.decode('utf-8')
 
 def whether_job_is_pending(state):
     """
@@ -74,10 +31,8 @@ def whether_job_is_pending(state):
     :return: true if a job state is not in finished state
     """
     status = ["pend", "pending", "pd", "configuring", "cf"]
-    if state.lower() in status:
-        return True
-    else:
-        return False
+    return state.lower() in status
+
 
 def whether_job_is_suspending(state):
     """
@@ -95,10 +50,8 @@ def whether_job_is_suspending(state):
     """
     status = ["psusp", "ususp", "ssusp", "prov", "wait", "suspended", "s", "requeued", "rq",
               "requeue_hold", "rh", "resv_del_hold", "rd", "requeue_fed", "rf"]
-    if state.lower() in status:
-        return True
-    else:
-        return False
+    return state.lower() in status
+
 
 def whether_job_is_running(state):
     """
@@ -115,10 +68,8 @@ def whether_job_is_running(state):
     :return: true if a job state is not in finished state
     """
     status = state.lower()
-    if status == "r" or status.find("run") >= 0:
-        return True
-    else:
-        return False
+    return status == "r" or status.find("run") >= 0
+
 
 def whether_job_is_finished(state):
     """
@@ -128,10 +79,10 @@ def whether_job_is_finished(state):
     :param state: string that representing a job state
     :return: true if a job state is not in finished state
     """
-    if whether_job_is_running(state) or whether_job_is_pending(state) or whether_job_is_suspending(state):
-        return False
-    else:
-        return True
+    return not (whether_job_is_running(state) or
+                whether_job_is_pending(state) or
+                whether_job_is_suspending(state))
+
 
 def get_time_data_from_lsf_output(data):
     """
@@ -212,43 +163,6 @@ def get_time_data_from_lsf_output(data):
     # finally return
     return t
 
-def convert_percentage_to_decimal(input):
-    """
-    convert the input percentage string into a decimal
-    :param input: number in string with percentage sign
-    :return: a decimal
-    """
-    # if input is empty, return 0
-    if not input.strip():
-        return 0.0
-
-    # if we have something, then percentage sign should be there
-    if input.find("%") < 0:
-        info = "Failed to find the percentage sign % in the input string for convert"
-        raise RuntimeError(info)
-
-    # now everything good, let's do the job
-    decimal_value = float(input.strip('%')) / 100
-    return decimal_value
-
-def convert_str_to_integer(input):
-    """
-    a simple function to convert the input string to integer number
-    :param input: input string, should be all digit
-    :return: an integer number
-    """
-    # if input is empty, return 0
-    if not input.strip():
-        return 0
-
-    # convert
-    s = input.strip()
-    if s.isdigit():
-        num = int(s)
-    else:
-        info = "The input string is not numeric: {}", input
-        raise RuntimeError(info)
-    return num
 
 def convert_lsf_time_to_minutes(input):
     """
@@ -272,7 +186,8 @@ def convert_lsf_time_to_minutes(input):
         raise RuntimeError(info)
 
     # now return the number
-    return v0*60 + v1
+    return v0 * 60 + v1
+
 
 def get_hostname_from_bjobs_output(input):
     """
@@ -286,50 +201,8 @@ def get_hostname_from_bjobs_output(input):
 
     # the string in format like 3*nodecem123, so node name is nodecem123
     s = input.strip()
-    if s.find('*') >=0:
+    if s.find('*') >= 0:
         return s.split('*')[1]
     else:
         info = "Invalid string for paring to get hostname: {}".format(input)
         raise RuntimeError(info)
-
-def get_gpu_type_for_node_from_lsf(input):
-    """
-    check the gpu type from the input
-    :param input: input string obtained from lsf command
-    :return: the gpu type
-    """
-    for t in GPU_TYPE:
-        t0 = t.split("_")[0]
-        mem_label = t.split("_")[1]
-
-        # if the type match
-        if input.find(t0) > 0 and input.find(mem_label) > 0:
-            return t
-
-    # now if we are here, that means we did not find anything match
-    info = "Invalid input GPU type for paring: {}".format(input)
-    raise RuntimeError(info)
-
-def read_data_from_csv(file_name):
-    """
-    reading data from the input csv file
-
-    we note, that the csv file must have header line; and should follow the example below:
-
-    x, y, z,  w (header line)
-    a1, a2, a3, a4
-    b1, b2, b3, b4
-    .........
-
-    the result will be a list of dict, for example
-    [ {x: a1, y: a2, z: a3, w: a4}, {x: b1, y: b2, z: b3, w: b4} ... ]
-
-    :param file_name: input csv file name
-    :return: a list of dict
-    """
-    data_list = []  # List to store dictionaries
-    with open(file_name, mode='r') as file:
-        csv_reader = csv.DictReader(file)  # Create DictReader
-        for row in csv_reader:
-            data_list.append(row)
-    return data_list
