+++ emgoat/process/relion_job_template.py
@@ -1,0 +1,7 @@
+
+OUTFILE = "XXXoutfileXXX"
+ERRFILE = "XXXerrfileXXX"
+
+COMMAND = """
+XXXcommandXXX
+"""
