+++ emgoat/scripts/emgoat.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+. /hpcf/authorized_apps/cryo_apps/ctools/bashrc
+
+python /home/jdela80/cryo_apps/tmp/stjude-biohackthon-2025/emgoat.py $@
+
