+++ emgoat/util/__init__.py
@@ -1,0 +1,2 @@
+from .macros import *
+from .util import *
