--- lsf/__init__.py
@@ -1,1 +1,0 @@
-from lsf.__version__ import __version__, __title__
