--- lsf/__version__.py
@@ -1,3 +1,0 @@
-__version__ = '0.1'
-
-__title__ = "LSF wrapper functions"
